requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==4.0.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
faiss-cpu==1.11.0.post1
filelock==3.18.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pillow==11.3.0
pinecone
pinecone-plugin-interface==0.0.7
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers
torch
torchaudio
torchvision
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
